Moved add_progressionsfaktor.
